Moves all VMHook related abstractions to separate package.
