+++ ethereumj-core/src/main/java/org/ethereum/vm/hook/VMHookFactory.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.vm;
+package org.ethereum.vm.hook;
 
 /**
  * Created by Richard Hart on 07.07.2018.
